testv1/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
aiohttp==3.7.4.post0
aiomysql==0.0.21
lxml==4.6.3
beautifulsoup4==4.9.3
python-dotenv==0.19.0
